Added CSV SPARQL results parser.
CSV parsing is by definition unreliable since the serialization format intentionally discards type information (URI/literal). Strings starting with http:// or https:// are parsed as URIs always (even if literals in the source) and all other values which are not blank nodes (starting with _:) are treated as plain string literals.

Collateral changes:

- AbstractSVResultsParser generalized out of TSVParser
- CSUtils.skipUntil{,In} with three chars generalized to char[]