Fix starvation on second NettySparqlClient.query().
The issue was that sending a DefaultHttpRequest instead of a DefaultFullHttpRequest left the netty HTTP encoder expecting a LastHttpContent before returning to `ST_INIT` state. Since the encoder never returned to the `ST_INIT` state, an attempt to encode a new request after the Channel was recycled via the pool failed with an IllegalStateException.