Overhaul of bind and ws protocol.
Having the server echo all bindings back to the client is wasteful, thus the WS server now sends a base64-encoded unsigned 48 bit sequence number for each result indicating the sequence number of the binding that yielded those results. For EXISTS/MINUS binds, the results serialization consists entirely of sequence numbers.

Due to the use of sequence numbers, all bind implementations had to be changed to allow NettySparqlServer to keep track of how many bindings have been already consumed and up to which point the bindings yielded no results. This is necessary in order for the server to send the client a !bind-empty-streak N command to the client together with binding requests that enables the client to remove queue bindings up to the sequence number N. This is necessary to avoid having the client's sendBindingsThread starve while adding to the sentBindings queue.